% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paging_functions.R
\name{add_type}
\alias{add_type}
\title{Function comparing columns of two dataframes and adding missing columns where needed.}
\usage{
add_type(df1, df2, verbose = FALSE)
}
\arguments{
\item{df1}{(dataframe) First dataframe}

\item{df2}{(dataframe) Second dataframe}
}
\value{
list with two dataframes
}
\description{
This function adds columns to two dataframes based on columns of both
}
