% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paging_functions.R
\name{check_type}
\alias{check_type}
\title{Function checking columns types of two dataframes}
\usage{
check_type(df1, df2)
}
\arguments{
\item{v1}{(dataframe) First dataframe}

\item{v2}{(dataframe) Second dataframe}
}
\value{
A logical vector
}
\description{
This function checks types of corresponding columns from two dataframes.
TRUE if they are different, FALSE otherwise
}
