% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_import.R
\name{checklist_import}
\alias{checklist_import}
\title{Import GBIF or ad hoc checklist}
\usage{
checklist_import(checklist, tot_limit = NULL)
}
\arguments{
\item{checklist}{GBIF checklist datasetKey or tsv file path}

\item{tot_limit}{Maximum number of records to return for each dataset. 
Default NULL, that means all records of every dataset}
}
\value{
A list with two elements: 
species, list with data from checklists
}
\description{
This function opens a checklist from GBIF and/or from a tsv file as R dataframes.
}
