% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_import.R
\name{checklists_import}
\alias{checklists_import}
\title{Import GBIF or ad hoc checklists}
\usage{
checklists_import(checklists, tot_limit = NULL, verbose = FALSE)
}
\arguments{
\item{checklists}{list containing all GBIF checklist datasetKey or tsv file paths}

\item{tot_limit}{number of records to import from each dataset. 
This function supports only default valuue NULL, that means importing all records.}
}
\value{
dataframe
}
\description{
This function calls checklist_import for opening all checklists
}
