% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_shapes}
\alias{climate_shapes}
\alias{observed}
\alias{future}
\alias{legend}
\title{Climate shapefiles}
\format{
\code{observed} is a list of 5 spatialpolygonsdataframes:
\itemize{
\item \code{1901-1925}: observed climate data from 1901 up to 1925
\item \code{1925-1950}: observed climate data from 1926 up to 1950
\item \code{1950- 1975}: observed climate data from 1951 up to 1975
\item \code{1976-2000}: observed climate data from 1976 up to 2000
\item \code{1980-2016}: observed climate data from 1980 up to 2016
}

Each spatialpolygonsdataframe contains 2 variables:
\itemize{
\item \code{ID}: polygon identifier
\item \code{GRIDCODE}: grid value corresponding to a climate zone
}

\code{future} is a list of 5 spatialpolygonsdataframes:
\itemize{
\item \code{2001-2025-A1F1}: A1FI scenario for the possible climate between 2001 and 2025
\item \code{2026-2050-A1FI}: A1FI scenario for the possible climate between 2026 and 2050
\item \code{2051-2075-A1FI}: A1FI scenario for the possible climate between 2051 and 2075
\item \code{2076-2100-A1FI}: A1FI scenario for the possible climate between 2076 and 2100
\item \verb{2071-2100_Beck}: Beck scenario for the possible climate between 2071 and 2100
}

Each spatialpolygonsdataframe contains 2 variables:
\itemize{
\item \code{ID}: polygon identifier
\item \code{GRIDCODE}: grid value corresponding to a climate zone
}

\code{legend} contains two data.frames, \code{KG_A1FI} and \code{KG_Beck_Legend},
corresponding to A1FI and Beck scenarios respectively. Each data.fram
contains two columns:
\itemize{
\item \code{GRIDCODE}: grid value corresponding to a climate zone
\item \code{Classification}: climate classification value
}
}
\usage{
observed

future

legend
}
\description{
These shapefiles contain worldwide climate classifications for a
given period and are grouped according to their period.

Historical/observed climate classification shapes are grouped in the
\code{observed} datapackage. These shapes all originate from \href{http://dx.doi.org/10.1127/0941-2948/2010/0430}{Rubel & Kottek 2010}. Future scenarios are
dependant on several variables like pollution levels. \href{http://dx.doi.org/10.1127/0941-2948/2010/0430}{Rubel & Kottek 2010.} and \href{https://doi.org/10.1038/sdata.2018.214}{Beck et al. 2018} describe several scenarios for
the possible future climates. Currently the \code{future} datapackage contains the
following scenarios:
\itemize{
\item A1FI: from Rubel & Kottek 2010. Quick economic and technological growth
through intensive use of fossil fuel
\item Beck: from Beck et al. 2018. high emissions
}

Legends are stored in the \code{legend} datapackage
}
\concept{climate_shapefiles}
\keyword{datasets}
