% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paging_functions.R
\name{coerce_datatypes}
\alias{coerce_datatypes}
\title{Function containing binary coercions rules}
\usage{
coerce_datatypes(df1, df2, diff_type)
}
\arguments{
\item{df1}{(dataframe) First dataframe}

\item{df2}{(dataframe) Second dataframe}

\item{diff_type}{(vector) TRUE for columns where type coercion is needed}
}
\value{
A list with two dataframes
}
\description{
This function contains coercion rules and will be constantly updated.
}
