% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paging_functions.R
\name{coerce_table}
\alias{coerce_table}
\title{Function containing binary coercions rules}
\usage{
coerce_table(v1, v2)
}
\arguments{
\item{v1}{(vector) First vector}

\item{v2}{(vector) Second vector}
}
\value{
A vector containing the modified vectors
}
\description{
This function contains coercion rules and will be constantly updated.
}
