% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-future.R
\docType{data}
\name{future}
\alias{future}
\title{Future climate list of sf objects}
\format{
\code{future} is a list of 5 sf objects:
\itemize{
\item \code{2001-2025-A1F1}: A1FI scenario for the possible climate between 2001 and 2025
\item \code{2026-2050-A1FI}: A1FI scenario for the possible climate between 2026 and 2050
\item \code{2051-2075-A1FI}: A1FI scenario for the possible climate between 2051 and 2075
\item \code{2076-2100-A1FI}: A1FI scenario for the possible climate between 2076 and 2100
\item \verb{2071-2100_Beck}: Beck scenario for the possible climate between 2071 and 2100
}

Each sf object contains 3 variables:
\itemize{
\item \code{ID}: polygon identifier
\item \code{GRIDCODE}: grid value corresponding to a climate zone
\item \code{geometry}: the coordinates that define the polygon's shape
}
}
\source{
\href{http://dx.doi.org/10.1127/0941-2948/2010/0430}{Rubel & Kottek 2010} and
\href{https://doi.org/10.1038/sdata.2018.214}{Beck et al. 2018}.
}
\usage{
future
}
\description{
These sf objects contain future worldwide climate
classifications for different year intervals.

Future scenarios are dependent on several variables like pollution levels.
Currently the \code{future}
datapackage contains the following scenarios:
\itemize{
\item A1FI: from Rubel & Kottek 2010, quick economic and technological growth
through intensive use of fossil fuel
\item Beck: from Beck et al. 2018, high emissions
}
}
\seealso{
Other climate data:
\code{\link{legends}},
\code{\link{observed}}
}
\concept{climate data}
\keyword{datasets}
