% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_has_distribution.R
\docType{data}
\name{GBIF_distr_terms}
\alias{GBIF_distr_terms}
\title{Accepted GBIF distribution terms}
\format{
An object of class \code{character} of length 3.
}
\usage{
GBIF_distr_terms
}
\description{
Accepted GBIF distribution terms
}
\keyword{internal}
