% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_download.R
\name{gbif_download}
\alias{gbif_download}
\title{Trigger a download from GBIF}
\usage{

  gbif_download(taxa = "https://raw.githubusercontent.com/trias-project/alien-plants-belgium/afbd2805de77afd79fb74669c403d40f1416661b/data/processed/taxon.csv",
  country = "BE", output = "./data/output/gbif_downloads.csv")
}
\arguments{
\item{taxa}{Path to a csv file with informations about species}

\item{country}{Country (default: BE)}

\item{output}{Path to output csv file}
}
\description{
This function uses GBIF to trigger a download of a set of  features for a list of species.
}
