% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_taxon_infos_extraction.R
\name{gbif_taxon_infos_extraction}
\alias{gbif_taxon_infos_extraction}
\title{Add informations provided by the GBIF taxonomy informations}
\usage{
gbif_taxon_infos_extraction(nubKey_info, nubKey_col, gbif_terms)
}
\arguments{
\item{nubKey_info}{list with informations about nubKeys}

\item{nubKey_col}{char column name of the column containing the nubKeys
used for retrieving additional information.}

\item{gbif_terms}{list of valid GBIF terms to add as additional information}
}
\value{
list with GBIF information as additional columns
}
\description{
This function extends an existing dataframe with additional columns provided 
by the GBIF taxonomic backbone and matched on the nubkey,
which needs to be an available column in the dataframe.
}
\details{
This function is essentially a wrapper around the existing rgbif
functionality `name_usage` and extends the application to a data.frame.
For more information on the name matching API of GBIF on which rgbif relies,
see https://www.gbif.org/developer/species
}
