% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_verify_synonyms.R
\name{gbif_verify_synonyms}
\alias{gbif_verify_synonyms}
\title{Check and update verified synonyms}
\usage{
gbif_verify_synonyms(taxa, verified_synonyms)
}
\arguments{
\item{taxa:}{a dataframe with at least the following columns: \itemize{
\item{backbone_taxonKey} \item{backbone_scientificName}
\item{backbone_acceptedKey} \item{backbone_accepted} }}

\item{verified_synonyms:}{a dataframe with at least the following columns:
\itemize{ \item{backbone_taxonKey} \item{backbone_scientificName}
\item{backbone_acceptedKey} \item{backbone_accepted}
\item{backbone_kingdom}{: to be populated from GBIF (is not in taxa)}
\item{date_added}{: to be populated by function} \item{verified_key}{: to
be populated manually by expert (not required by this function, but any
other functionality will use this key so it is good to check its existence)
} }}
}
\value{
a list of five dataframes: \itemize{ \item{verified_synonyms}{: same
  dataframe as input verified_synonyms, but now with updated info.}
  \item{new_synonyms}{: a subset of verified_synonyms (same columns) with
  added synonym relations (found in taxa, but not in verified_synonyms)}
  \item{unused_synonyms}{: a subset of verified_synonyms (same columns) with
  unused synonym relations (found in verified_synonyms, but not in taxa)}
  \item{updated_scientificName}{: a dataframe with backbone_scientificName +
  updated_backbone_scientificName} \item{updated_accepted}{: a dataframe with
  backbone_accepted + updated_backbone_accepted} }
}
\description{
This function checks all verified synonyms, add new sysnonyms in order to be
evaluated by an expert, update taxa names in case they have been changed and
report the changes.
}
\examples{
taxa_in <- data.frame(
  backbone_taxonKey = c(2360181, 2427092, 2651108),
  backbone_scientificName = c("Aspius aspius (Linnaeus, 1758)",
                              "Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-simense (Hook.) J.Sm."),
  backbone_acceptedKey = c(5851603, 2427091, 4046493),
  backbone_accepted = c("Leuciscus aspius (Linnaeus, 1758)",
                       "Lithobates catesbeianus (Shaw, 1802)",
                       "Polystichum luctuosum (Kunze) Moore."),
  stringsAsFactors = FALSE)
verified_synonyms_in <- data.frame(
  backbone_taxonKey = c(2427092,
                        2651108,
                        6723),
  backbone_scientificName = c("Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-tsus-tsus (Hook.) Captain",
                              "Lemnaceae"),
  backbone_accepted = c("Lithobates dummyus (Batman, 2018)",
                        "Polystichum luctuosum (Kunze) Moore.",
                        "Araceae"),
  backbone_acceptedKey = c(2427091,
                           4046493,
                           6979),
  backbone_kingdom = c("Animalia",
                       "Plantae",
                       "Plantae"),
  date_added = as.Date(c("2018-01-01",
                         "2018-01-01",
                         "2018-01-01")),
  verified_key = c(2427091,
                   4046493,
                   6979),
  remarks = c("dummy example 1: backbone_accepted should be updated",
              "dummy example 2: backbone_scientificName should be updated",
              "dummy example 3: nothing should be changed"),
  stringsAsFactors = FALSE)
gbif_verify_synonyms(taxa = taxa_in, verified_synonyms = verified_synonyms_in)
}
