% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cred.R
\name{get_cred}
\alias{get_cred}
\title{Interactively get & store credentials in the system environment}
\usage{
get_cred(x)
}
\arguments{
\item{x}{a character with the name of the system environment variable to get
and store}
}
\value{
a character vector containing the value of the system variable
}
\description{
Interactively get & store credentials in the system environment
}
