% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_pathways.R
\name{get_table_pathways}
\alias{get_table_pathways}
\title{Pathway count indicator table}
\usage{
get_table_pathways(
  df,
  category = NULL,
  from = NULL,
  n_species = 5,
  kingdom_names = "kingdom",
  phylum_names = "phylum",
  first_observed = "first_observed",
  species_names = "canonicalName"
)
}
\arguments{
\item{df}{df.}

\item{category}{NULL or character. One of the kingdoms as given in GBIF:
\itemize{ \item{"Plantae"} \item{"Animalia"} \item{"Fungi"}
\item{"Chromista"} \item{"Archaea"} \item{"Bacteria"} \item{"Protozoa"}
\item{"Viruses"} \item{"incertae sedis"} } It can also be one of the
following not kingdoms: #'\itemize{\item{Chordata} \item{Not Chordata}.
Default: NULL.}}

\item{from}{NULL or numeric. Year trade-off: if not \code{NULL} select only
pathways related to taxa introduced during or after this year. Default:
\code{NULL}.}

\item{n_species}{numeric. The maximum number of species to return as examples
per pathway. For groups with less species than \code{n_species}, all
species are given. Default: 5.}

\item{kingdom_names}{character. Name of the column of \code{df} containing
information about kingdom. Default: \code{"kingdom"}.}

\item{phylum_names}{character. Name of the column of \code{df} containing
information about phylum. This parameter is used only if \code{category} is
one of:  \code{"Chordata"}, \code{"Not Chordata"}.  Default: \code{"phylum"}.}

\item{first_observed}{character. Name of the column of \code{df}
containing information about year of introduction. Default:
\code{"first_observed"}.}

\item{species_names}{character. Name of the column of \code{df} containing
information about species names. Default: \code{"canonicalName"}.}
}
\value{
a data.frame with 4 columns: \code{pathway_level1}, \code{pathway_level2}, \code{n}
(number of taxa) and \code{examples}.
}
\description{
Function to get number of taxa introduced by different pathways. Possible
breakpoints: taxonomic (kingdom + vertebrates/invertebrates), temporal (lower
limit year).
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/indicators/master/data/",
  "interim/data_input_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "NA",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    acceptedKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
get_table_pathways(data)
# Specify kingdom
get_table_pathways(data, "Plantae")
# with special categories, `Chordata` or `not Chordata`
get_table_pathways(data, "Chordata")
get_table_pathways(data, "Not Chordata")
# From 2000
get_table_pathways(data, from = 2000, first_observed = "first_observed")
# Specify number of species to include in examples
get_table_pathways(data, "Plantae", n_species = 8)
# Specify columns containing kingdom and species names
get_table_pathways(data,
  "Plantae",
  n_species = 8,
  kingdom_names = "kingdom",
  species_names = "canonicalName"
)
}
}
\seealso{
Other checklist functions:
\code{\link[=gbif_get_taxa]{gbif_get_taxa()}},
\code{\link[=gbif_has_distribution]{gbif_has_distribution()}},
\code{\link[=gbif_verify_keys]{gbif_verify_keys()}},
\code{\link[=get_nubkeys]{get_nubkeys()}},
\code{\link[=indicator_introduction_year]{indicator_introduction_year()}},
\code{\link[=indicator_native_range_year]{indicator_native_range_year()}},
\code{\link[=indicator_total_year]{indicator_total_year()}},
\code{\link[=pathways_cbd]{pathways_cbd()}},
\code{\link[=verify_taxa]{verify_taxa()}},
\code{\link[=visualize_pathways_level1]{visualize_pathways_level1()}},
\code{\link[=visualize_pathways_level2]{visualize_pathways_level2()}},
\code{\link[=visualize_pathways_year_level1]{visualize_pathways_year_level1()}},
\code{\link[=visualize_pathways_year_level2]{visualize_pathways_year_level2()}}
}
\concept{checklist functions}
