% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_introduction_year.R
\name{indicator_introduction_year}
\alias{indicator_introduction_year}
\title{Plot the number of new introductions per year}
\usage{
indicator_introduction_year(
  df,
  start_year_plot = 1920,
  smooth_span = 0.85,
  x_major_scale_stepsize = 10,
  x_minor_scale_stepsize = 5,
  facet_column = NULL,
  taxon_key_col = "key",
  first_observed = "first_observed",
  x_lab = "Year",
  y_lab = "Number of introduced alien species"
)
}
\arguments{
\item{df}{A data frame.}

\item{start_year_plot}{Year where the plot starts from. Default: 1920.}

\item{smooth_span}{(numeric) Parameter for the applied
\code{\link[stats]{loess}} smoother. For more information on the
appropriate value, see \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Default: 0.85.}

\item{x_major_scale_stepsize}{(integer) Parameter that indicates the breaks
of the x axis. Default: 10.}

\item{x_minor_scale_stepsize}{(integer) Parameter that indicates the minor
breaks of the x axis. Default: 5.}

\item{facet_column}{NULL or character. The column to use to create additional
facet wrap plots underneath the main graph. When NULL, no facet graph are
created. Valid facet options: \code{"family"}, \code{"order"},
\code{"class"}, \code{"phylum"}, \code{"kingdom"}, \code{"pathway_level1"},
\code{"locality"}, \code{"native_range"} or  \code{"habitat"}. Default:
NULL.}

\item{taxon_key_col}{character. Name of the column of \code{df} containing
unique taxon IDs. Default: \code{key}.}

\item{first_observed}{character. Name of the column of \code{df} containing
information about year of introduction. Default: \code{first_observed}.}

\item{x_lab}{NULL or character. to set or remove the x-axis label.}

\item{y_lab}{NULL or character. to set or remove the y-axis label.}
}
\value{
A list with three slots:
\itemize{
\item \code{plot}: ggplot2 object (or egg object if facets are used).
\item \code{data_top_graph}: data.frame (tibble) with data used for the main plot (top graph) in \code{plot}.
\item \code{data_facet_graph}: data.frame (tibble) with data used for the faceting
plot in \code{plot}. If \code{facet_column} is NULL, NULL is returned.
}
}
\description{
Calculate how many new species has been introduced in a year.
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/indicators/master/data/",
  "interim/data_input_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
# without facets
indicator_introduction_year(data)
# specify start year and smoother parameter
indicator_introduction_year(data,
  start_year_plot = 1940,
  smooth_span = 0.6
)
# with facets
indicator_introduction_year(data, facet_column = "kingdom")
# specify columns with year of first observed
indicator_introduction_year(data,
  first_observed = "first_observed"
)
# specify axis labels
indicator_introduction_year(data, x_lab = "YEAR", y_lab = NULL)
}
}
\seealso{
Other checklist functions:
\code{\link[=gbif_get_taxa]{gbif_get_taxa()}},
\code{\link[=gbif_has_distribution]{gbif_has_distribution()}},
\code{\link[=gbif_verify_keys]{gbif_verify_keys()}},
\code{\link[=get_nubkeys]{get_nubkeys()}},
\code{\link[=get_table_pathways]{get_table_pathways()}},
\code{\link[=indicator_native_range_year]{indicator_native_range_year()}},
\code{\link[=indicator_total_year]{indicator_total_year()}},
\code{\link[=pathways_cbd]{pathways_cbd()}},
\code{\link[=verify_taxa]{verify_taxa()}},
\code{\link[=visualize_pathways_level1]{visualize_pathways_level1()}},
\code{\link[=visualize_pathways_level2]{visualize_pathways_level2()}},
\code{\link[=visualize_pathways_year_level1]{visualize_pathways_year_level1()}},
\code{\link[=visualize_pathways_year_level2]{visualize_pathways_year_level2()}}
}
\concept{checklist functions}
