% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_pathway.R
\name{indicator_pathway}
\alias{indicator_pathway}
\title{Pathways of introduction}
\usage{
indicator_pathway(data, subcategory = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{subcategory}{(boolean) If TRUE, the subcategory levels are included in
the visualisation.}
}
\value{
A ggplot2 figure.
}
\description{
Make a figure of pathways of introduction. Use  \code{subcategory}
option to take into account only main categories or include subcategories as
well.
}
\examples{
\dontrun{
indicator_pathway(data)
indicator_pathway(data, subcategory = TRUE)
}
}
