% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-legend.R
\docType{data}
\name{legend}
\alias{legend}
\title{Legends for climate shapefiles}
\format{
\code{legend} contains two data.frames, \code{KG_A1FI} and \code{KG_Beck_Legend},
matching Koppen-Geiger climate zones to A1FI and Beck scenarios
respectively.

Each data.frame contains two columns:
\itemize{
\item \code{GRIDCODE}: (numeric) grid value corresponding to a climate zone
\item \code{Classification}: (character) Koppen–Geiger climate classification value
}
}
\source{
\href{https://hess.copernicus.org/articles/11/1633/2007/}{Koppen-Geiger climate zones}
}
\usage{
legend
}
\description{
Legends for climate shapefiles
}
\seealso{
Other climate_shapefiles: 
\code{\link{future}},
\code{\link{observed}}
}
\concept{climate_shapefiles}
\keyword{datasets}
