% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-legends.R
\docType{data}
\name{legends}
\alias{legends}
\title{Legends for climate shapefiles}
\format{
\code{legends} contains two data.frames, \code{KG_A1FI} and \code{KG_Beck},
matching Koppen-Geiger climate zones to A1FI and Beck scenarios
respectively.

Each data.frame contains two columns:
\itemize{
\item \code{GRIDCODE}: (numeric) grid value corresponding to a climate zone
\item \code{Classification}: (character) Koppen-Geiger climate classification value
\item \code{Description}: (character) verbose description of the Koppen-Geiger
climate zone, e.g. "Tropical rainforest climate"
\item \code{Group}: (character) group the Koppen-Geiger climate zone belongs to,
e.g. "Tropical"
\item \verb{Precipitation Type}: (character) Type of precipitations associated to
the climate zone, e.g. "Rainforest"
\item \verb{Level of Heat}: (character) Heat level associated to the climate zone,
e.g. "Cold"
}
}
\source{
\href{https://hess.copernicus.org/articles/11/1633/2007/}{Koppen-Geiger climate zones}
}
\usage{
legends
}
\description{
Legends for climate shapefiles
}
\seealso{
Other climate data:
\code{\link{future}},
\code{\link{observed}}
}
\concept{climate data}
\keyword{datasets}
