% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_usage_paging.R
\name{name_usage_paging}
\alias{name_usage_paging}
\title{Retrieve all records contained in a GBIF checklist}
\usage{
name_usage_paging(datasetKey, data = "all", offset = 0, tot_limit = NULL,
  verbose = FALSE)
}
\arguments{
\item{datasetKey}{(character) Filters by the GBIF dataset’s key (a uuid)}

\item{data}{(character) Specify an option to select what data is returned. 
See name_usage documentation.}

\item{offset}{Record number to start at. Default: 0 (as in name_usage) 
Use in combination with limit to page through results}

\item{tot_limit}{Number of records to return. 
Default NULL (return all records)}
}
\value{
A list of length two. 
The first element is metadata. The second is a data.frame (as in name_usage)
}
\description{
This function retrieves all records contained in a GBIF checklist.
It implements paging by using iteratively rgbif fucntion name_usage.
}
