% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_introduction_year.R
\name{nice_seq}
\alias{nice_seq}
\title{Generate nice sequence with slightly different cut values compared to \code{\link[base]{seq}}}
\usage{
nice_seq(start_year, end_year, step_size)
}
\arguments{
\item{start_year}{(integer) The min cut value}

\item{end_year}{(integer) The max cut value}

\item{step_size}{(integer) The max distance between two cut values}
}
\value{
integer vector, all cut values
}
\description{
The inner values are forced to be a multiple of the stepsize
}
\author{
mvarewyck
}
