% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pathwayscbd.R
\docType{data}
\name{pathwayscbd}
\alias{pathwayscbd}
\title{Pathways of introduction as defined by CBD}
\format{
A tibble with 51 rows and 2 variables:
\itemize{
\item \code{pathway_level1}: (character) CBD pathway level 1
\item \code{pathway_level2}: (character) CBD pathway level 2
}
}
\source{
\href{https://www.cbd.int/}{CBD Standard}
}
\usage{
pathwayscbd
}
\description{
A tibble containing all CBD pathways of introduction at level 1
(\code{pathway_level1}) and level 2 (\code{pathway_level2}). Added pathway \code{unknown}
at level 1 and level 2 for classifying taxa without pathway (at level 1 or
level 2) information.
}
\concept{checklist data}
\keyword{datasets}
