% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_gam.R
\name{plot_ribbon_em}
\alias{plot_ribbon_em}
\title{Plot time series with confidence limits and emerging status}
\usage{
plot_ribbon_em(df_plot, x_axis = "year", y_axis = "obs",
  x_label = "x", y_label = "y", ptitle = NULL, verbose = FALSE)
}
\arguments{
\item{df_plot:}{df. A data.frame containing data to plot.}

\item{x_axis:}{character. Name of column containing x-values. Default:
\code{"year"}.}

\item{y_axis:}{character. Name of column containing y-values. Default:
\code{"number of observations"}.}

\item{x_label:}{character. x-axis label. Default: \code{"x"}.}

\item{y_label:}{character. y-axis label. Default: \code{"y"}.}

\item{ptitle:}{character. Plot title. Default: \code{NULL}.}

\item{verbose:}{logical. If \code{TRUE}, informations about possible issues
are returned. Default: \code{FALSE}.}
}
\value{
a ggplot2 plot object.
}
\description{
Plot time series with confidence limits and emerging status
}
