% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\title{Write a query as JSON object for downloading GBIF occurrences}
\usage{
predicates(taxonKey = taxonKeys, countries = countries, creator = NULL,
  notification_address = NULL)
}
\arguments{
\item{creator}{GBIF username}

\item{notification_address}{GBIF e-mail}

\item{taxonKeys}{vector with taxonKeys}

\item{country}{vector with countries}
}
\value{
body as JSON object
}
\description{
This function writes a body query (JSON object) for triggering 
a GBIF occurrence download.
}
