% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trias_utilities.R
\name{split_path}
\alias{split_path}
\title{Small function Utilities within trias package.}
\usage{
split_path(path)
}
\description{
Collection of short functions used in other R scripts within trias package.
}
