% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_with_duplicates.R
\name{spread_with_duplicates}
\alias{spread_with_duplicates}
\title{Spread a key-value pair across multiple columns in presence of duplicates}
\usage{
spread_with_duplicates(.df, key, value, by, fill = NA)
}
\arguments{
\item{.df}{A dataframe}

\item{key, value}{Column names or positions}

\item{fill}{If set, missing values will be replaced with this value.}
}
\value{
A data.frame
}
\description{
Spread a key-value pair across multiple columns in presence of duplicates
}
\examples{
test1 <- data.frame(taxonKey = c(1, 1, 1, 1),	
                key = c("A", "B", "C", "C"),
                value = c("R", "S", "T", "X"),
                stringsAsFactors = FALSE)
spread_with_duplicates(test1, key, value, by = "taxonKey")
# with NAs
test2 <- data.frame(taxonKey = c(1, 1, 1, 2),	
                key = c("A", "C", "C", "A"),
                value = c("R", "T", "X", "R"),
                stringsAsFactors = FALSE)
spread_with_duplicates(test2, key, value, by = "taxonKey")
spread_with_duplicates(test2, key, value, by = "taxonKey", fill = 0)
}
