% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pathways_level2.R
\name{visualize_pathways_level2}
\alias{visualize_pathways_level2}
\title{Plot number of introduced taxa for CBD pathways level 2}
\usage{
visualize_pathways_level2(
  df,
  chosen_pathway_level1,
  category = NULL,
  from = NULL,
  facet_column = NULL,
  pathways = NULL,
  pathway_level1_names = "pathway_level1",
  pathway_level2_names = "pathway_level2",
  taxon_names = "key",
  kingdom_names = "kingdom",
  phylum_names = "phylum",
  first_observed = "first_observed",
  cbd_standard = TRUE,
  title = NULL,
  x_lab = "Number of introduced taxa",
  y_lab = "Pathways"
)
}
\arguments{
\item{df}{df.}

\item{chosen_pathway_level1}{character. A pathway level 1. If CBD standard is
followed (see argument \code{cbd_standard}), one of the level 1 pathways from
\code{pathwayscbd}.}

\item{category}{\code{NULL} (default) or character. One of the kingdoms as given
in GBIF or \code{Chordata} (the phylum) or \verb{Not Chordata} (all other phyla of
\code{Animalia}):
\enumerate{
\item \code{Plantae}
\item \code{Animalia}
\item \code{Fungi}
\item \code{Chromista}
\item \code{Archaea}
\item \code{Bacteria}
\item \code{Protozoa}
\item \code{Viruses}
\item \verb{incertae sedis}
\item \code{Chordata}
\item \verb{Not Chordata}
}}

\item{from}{\code{NULL} or numeric. Year trade-off: if not \code{NULL} select only
pathways related to taxa introduced during or after this year. Default:
\code{NULL}.}

\item{facet_column}{\code{NULL} (default) or character. The column to use to
create additional facet wrap bar graphs underneath the main graph. When
\code{NULL}, no facet graph are created. One of \code{family}, \code{order}, \code{class},
\code{phylum}, \code{locality}, \code{native_range}, \code{habitat}. If column has another
name, rename it before calling this function. Facet \code{phylum} is not allowed
in combination with \code{category} equal to \code{"Chordata"} or \code{"Not Chordata"}.
Facet \code{kingdom} is allowed only with category equal to \code{NULL}.}

\item{pathways}{character. Vector with pathways level 2 to visualize. The
pathways are displayed following the order as in this vector. It may
contain pathways not present in the column given by \code{pathway_level1_names}.}

\item{pathway_level1_names}{character. Name of the column of \code{df}
containing information about pathways at level 1. Default:
\code{pathway_level1}.}

\item{pathway_level2_names}{character. Name of the column of \code{df}
containing information about pathways at level 2. Default:
\code{pathway_level2}.}

\item{taxon_names}{character. Name of the column of \code{df} containing
information about taxa. This parameter is used to uniquely identify taxa.}

\item{kingdom_names}{character. Name of the column of \code{df} containing
information about kingdom. Default: \code{"kingdom"}.}

\item{phylum_names}{character. Name of the column of \code{df} containing
information about phylum. This parameter is used only if \code{category} is
one of: \code{"Chordata"}, \code{"Not Chordata"}.  Default:
\code{"phylum"}.}

\item{first_observed}{character. Name of the column of \code{df} containing
information about year of introduction. Default: \code{"first_observed"}.}

\item{cbd_standard}{logical. If \code{TRUE} the values of pathway level 2 are
checked based on CBD standard as in \code{pathwayscbd}. Error is
returned if unmatched values are found. If \code{FALSE}, a warning is returned.
Default: \code{TRUE}.}

\item{title}{\code{NULL} or character. Title of the graph. Default: \code{NULL}.}

\item{x_lab}{\code{NULL} or character. x-axis label. Default: \code{"Number of introduced taxa"}.}

\item{y_lab}{\code{NULL} or character. Title of the graph. Default: \code{"Pathways"}.}
}
\value{
A list with three slots:
\itemize{
\item \code{plot}: ggplot2 object (or egg object if facets are used). \code{NULL} if there
are no data to plot.
\item \code{data_top_graph}: data.frame (tibble) with data used for the main plot (top
graph) in \code{plot}.
\item \code{data_facet_graph}: data.frame (tibble) with data used for the faceting
plot in \code{plot}. \code{NULL} is returned if \code{facet_column} is \code{NULL}.
}
}
\description{
Function to plot bar graph with number of taxa introduced by different
pathways at level 2, given a pathway level 1. Possible breakpoints: taxonomic
(kingdoms + vertebrates/invertebrates) and temporal (lower limit year).
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/indicators/master/data/",
  "interim/data_input_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
# All taxa
visualize_pathways_level2(data, chosen_pathway_level1 = "escape")

# Animalia
visualize_pathways_level2(data,
  chosen_pathway_level1 = "escape",
  category = "Animalia"
)

# Chordata
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  category = "Chordata"
)

# Select some pathways only
visualize_pathways_level2(
  df = data, 
  chosen_pathway_level1 = "escape",
  pathways = c("pet", "horticulture")
)

# `pathways` can contain pathways not present in data
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  category = "Chordata",
  pathways = c(
    "agriculture", # not present
    "forestry", # not present
    "pet", 
    "unknown"
  )
)

# facet phylum
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  category = "Animalia",
  facet_column = "phylum"
)

# facet habitat
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  facet_column = "habitat"
)

# Only taxa introduced from 1950
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  from = 1950
)

# Add a title
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  category = "Plantae",
  from = 1950,
  title = "Pathway level 2 (escape): Plantae, from 1950"
)

# Personalize axis labels
visualize_pathways_level2(
  df = data,
  chosen_pathway_level1 = "escape",
  x_lab = "Aantal taxa",
  y_lab = "pathways"
)
}
}
\seealso{
Other checklist functions:
\code{\link[=gbif_get_taxa]{gbif_get_taxa()}},
\code{\link[=gbif_has_distribution]{gbif_has_distribution()}},
\code{\link[=gbif_verify_keys]{gbif_verify_keys()}},
\code{\link[=get_nubkeys]{get_nubkeys()}},
\code{\link[=get_table_pathways]{get_table_pathways()}},
\code{\link[=indicator_introduction_year]{indicator_introduction_year()}},
\code{\link[=indicator_native_range_year]{indicator_native_range_year()}},
\code{\link[=indicator_total_year]{indicator_total_year()}},
\code{\link[=pathways_cbd]{pathways_cbd()}},
\code{\link[=verify_taxa]{verify_taxa()}},
\code{\link[=visualize_pathways_level1]{visualize_pathways_level1()}},
\code{\link[=visualize_pathways_year_level1]{visualize_pathways_year_level1()}},
\code{\link[=visualize_pathways_year_level2]{visualize_pathways_year_level2()}}
}
\concept{checklist functions}
