% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_decision_rules.R
\name{apply_decision_rules}
\alias{apply_decision_rules}
\title{Apply decision rules to time series and assess emerging status}
\usage{
apply_decision_rules(df, y_var = "ncells", eval_year, year = "year",
  taxonKey = "taxonKey")
}
\arguments{
\item{df}{df. A dataframe containing temporal data of one or more taxa. The
column with taxa can be of class character, numeric or integers.}

\item{y_var}{character. Name of column of \code{df} containing variable to
model. It has to be passed as string, e.g. \code{"occurrences"}.}

\item{eval_year}{numeric. Temporal value at which emerging status has to be
evaluated. \code{eval_year} should be present in timeseries of at least one
taxon.}

\item{year}{character. Name of column of \code{df} containing temporal
values. It has to be passed as string, e.g. \code{"time"}. Default:
\code{"year"}.}

\item{taxonKey}{character. Name of column of \code{df} containing taxon IDs.
It has to be passed as string, e.g. \code{"taxon"}. Default:
\code{"taxonKey"}.}
}
\value{
df. A dataframe (tibble) containing emerging status. Columns:
  \itemize{\item{\code{taxonKey}: column containing taxon ID. Column name
  equal to value of argument \code{taxonKey}.} \item{\code{year}: column
  containing temporal values. Column name equal to value of argument
  \code{year}. Column itself is equal to value of argument \code{eval_year}.
  So, if you apply decision rules on years 2018 (\code{eval_year = 2018}),
  you will get 2018 in this column.} \item{\code{em_status}: numeric.
  Emerging status, an integer between 0 and 3, based on output of decision
  rules (next columns).} \item{\code{dr_1}: logical. Output of decision rule
  1 answers to the question: does the time series contain only one positive
  value at evaluation year?} \item{\code{dr_2}: logical. Output of decision
  rule 2 answers to the question: is value at evaluation year above median
  value?} \item{\code{dr_3}: logical. Output of decision rule 3 answers to
  the question: does the time series contains only zeros in the five years
  before \code{eval_year}?} \item{\code{dr_4}: logical. Output of decision
  rule 4 answers to the question: is the value in column \code{y_var} the
  maximum ever observed up to \code{eval_year}?}}
}
\description{
This function defines and applies some decision rules to assess emerging
status at a specific time.
}
\examples{
\dontrun{
df <- tibble(
  taxonID = c(rep(1008955, 10), rep(2493598, 3)),
  y = c(seq(2009, 2018), seq(2016, 2018)),
  obs = c(1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 3, 0)
)
apply_decision_rules(df,
  eval_year = c(2016, 2017, 2018),
  y_var = "obs",
  taxonKey = "taxonID",
  year = y
)
}
}
