% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_has_distribution.R
\name{gbif_has_distribution}
\alias{gbif_has_distribution}
\title{Compare desired distribution information with actual one}
\usage{
gbif_has_distribution(taxon_key, ...)
}
\arguments{
\item{taxon_key}{(single numeric or character) a single taxon key.}

\item{...}{one or more GBIF distribution properties and related values.
Up to now it supports the following properties:
country (and its synonym: countryCode), status (and its synonym:
occurrenceStatus) and establishmentMeans.}
}
\value{
a logical, TRUE or FALSE.
}
\description{
This function compares GBIF distribution information based on a single taxon
key with user requests and returns a logical (TRUE or FALSE). Comparison is
case insensitive. User properties for each term are treated as OR.
It is a function built on rgbif function \code{name_usage()}.
}
\examples{
\dontrun{
# IMPORTANT! 
# examples could fail as long as `status` (`occurrenceStatus`) is used due to
# an issue of the GBIF API: see https://github.com/gbif/gbif-api/issues/94

# numeric taxonKey, atomic parameters
gbif_has_distribution(145953242,
  country = "BE",
  status = "PRESENT",
  establishmentMeans = "INTRODUCED"
)

# character taxonKey, distribution properties as vectors, treated as OR
gbif_has_distribution("145953242",
  country = c("NL", "BE"),
  status = c("PRESENT", "DOUBTFUL")
)

# use alternative names: countryCode, occurrenceStatus.
# Function works. Warning is given.
gbif_has_distribution("145953242",
  countryCode = c("NL", "BE"),
  occurrenceStatus = c("PRESENT", "DOUBTFUL")
)

# Case insensitive
gbif_has_distribution("145953242",
  country = "be",
  status = "PRESENT",
  establishmentMeans = "InTrOdUcEd"
)
}
}
\seealso{
Other checklist functions: 
\code{\link{gbif_get_taxa}()},
\code{\link{gbif_verify_keys}()},
\code{\link{get_table_pathways}()},
\code{\link{indicator_introduction_year}()},
\code{\link{indicator_native_range_year}()},
\code{\link{indicator_total_year}()},
\code{\link{pathways_cbd}()},
\code{\link{verify_taxa}()},
\code{\link{visualize_pathways_level1}()},
\code{\link{visualize_pathways_level2}()},
\code{\link{visualize_pathways_year_level1}()},
\code{\link{visualize_pathways_year_level2}()}
}
\concept{checklist functions}
