% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_has_distribution.R
\name{gbif_has_distribution}
\alias{gbif_has_distribution}
\title{Compare desired distribution information with actual one.}
\usage{
gbif_has_distribution(taxon_key, ...)
}
\arguments{
\item{taxon_key}{(single numeric or character) a single taxon key.}

\item{...}{one or more GBIF distribution properties and related values.
Up to now it supports the following properties:
country (and its synoynym: countryCode), status (and its synonym:
occurrenceStatus) and establishmentMeans.}
}
\value{
a logical, TRUE or FALSE.
}
\description{
This function compares GBIF distribution information based on a single taxon
key with user requests and returns a logical (TRUE or FALSE). Comparison is
case insensitive. User properties for each term ar treated as OR.
It is a function built on rgbif function `name_usage()`.
}
\examples{
# numeric taxonKey, atomic parameters
gbif_has_distribution(145953242,
  country = "BE",
  status = "PRESENT",
  establishmentMeans = "INTRODUCED"
)

# character taxonKey, distribution properties as vectors, treated as OR
gbif_has_distribution("145953242",
  country = c("NL", "BE"),
  status = c("PRESENT", "DOUBTFUL")
)

# use alternative names: countryCode, occurrenceStatus.
# Function works. Warning is given.
gbif_has_distribution("145953242",
  countryCode = c("NL", "BE"),
  occurrenceStatus = c("PRESENT", "DOUBTFUL")
)

# Case insensitive
gbif_has_distribution("145953242",
  countryCode = "be",
  status = "PRESENT",
  establishmentMeans = "InTrOdUcEd"
)
}
