% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_verify_keys.R
\name{gbif_verify_keys}
\alias{gbif_verify_keys}
\title{Check keys against GBIF Backbone Taxonomy}
\usage{
gbif_verify_keys(keys)
}
\arguments{
\item{keys:}{a vector, a list, or a data.frame with column \code{key}.}
}
\value{
a dataframe with the following columns: \itemize{\item{\code{key}}{:
  (numeric) invalid keys} \item{\code{is_taxonKey}} {: (logical) is the key
  a valid GBIF taxon key?} \item{\code{is_from_gbif_backbone}} {: (logical)
  is the key a valid taxon key from GBIF Backbone Taxonomy checklist?}
  \item{\code{is_synonym}} {: (logical) is the key related to a synonym (not
  \code{ACCEPTED} or \code{DOUBTFUL})?}}
}
\description{
This function performs three checks: \itemize{\item{\code{keys} are valid
GBIF taxon keys (https://www.gbif.org/species/key_value returns a page
related to a taxa)} \item{\code{keys} are taxon keys of the GBIF Backbone
Taxonomy checklist (https://www.gbif.org/species/key_value returns a page
related to a taxa of the GBIF Backbone)} \item{\code{keys} are synonyms
(neither \code{ACCEPTED} nor \code{DOUBTFUL})}}.
}
\examples{

}
