% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nubkeys.R
\name{get_nubkeys}
\alias{get_nubkeys}
\title{Retrieve unique GBIF Backbone taxon keys from a species checklist}
\usage{
get_nubkeys(datasetKey, allow_synonyms = TRUE)
}
\arguments{
\item{datasetKey}{(character) Unique identifier of a GBIF species checklist.}

\item{allow_synonyms}{(logical) Default: TRUE. If \code{FALSE}, the accepted taxa
the synonyms refers to are returned instead of the synonyms themselves.}
}
\value{
A (unique) vector of GBIF Backbone taxon keys (nubKeys). If
\code{allow_synonyms} is \code{TRUE}, the keys are retrieved from
\code{rgbif::name_usage()$data} directly. If \code{allow_synonyms} is \code{FALSE}, the
accepted taxa keys are retrieved by calling \code{rgbif::name_usage()} for each
synonym key obtained from
\code{rgbif::name_usage()$data} and extracting the accepted taxa keys from
them. The final keys are returned as unique values.
}
\description{
This function is a wrapper around \code{rgbif::name_usage()} to retrieve the (unique) GBIF
Backbone taxon keys (nubKeys) from a species checklist identified by its
datasetKey. It allows to choose whether to return synonyms or accepted taxa
only. In this case, \code{rgbif::name_usage()} is used a second time to get all
the taxonomic information of the accepted taxa the synonyms are pointing to within the GBIF Backbone. Synonym relationships within the checklist itself are not taken into account.
}
\examples{
dataset_key <- "e082b10e-476f-43c1-aa61-f8d92f33029a"
# Retrieve all GBIF Backbone taxon keys: synonyms are included
get_nubkeys(datasetKey = dataset_key)
# Retrieve accepted taxa GBIF Backbone taxon keys
get_nubkeys(datasetKey = dataset_key, allow_synonyms = FALSE)
}
\seealso{
Other checklist functions: 
\code{\link{gbif_get_taxa}()},
\code{\link{gbif_has_distribution}()},
\code{\link{gbif_verify_keys}()},
\code{\link{get_table_pathways}()},
\code{\link{indicator_introduction_year}()},
\code{\link{indicator_native_range_year}()},
\code{\link{indicator_total_year}()},
\code{\link{pathways_cbd}()},
\code{\link{verify_taxa}()},
\code{\link{visualize_pathways_level1}()},
\code{\link{visualize_pathways_level2}()},
\code{\link{visualize_pathways_year_level1}()},
\code{\link{visualize_pathways_year_level2}()}
}
\concept{checklist functions}
