% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_pathways.R
\name{get_table_pathways}
\alias{get_table_pathways}
\title{Pathway count indicator figure}
\usage{
get_table_pathways(df, category = NULL, n_species = 5,
  kingdom = "kingdom")
}
\arguments{
\item{df}{df.}

\item{category}{NULL or character. One of the kingdoms as given in GBIF: \itemize{
\item{"Plantae"} \item{"Animalia"} \item{} \item{"Fungi"} \item{"Chromista"}
\item{"Archaea"} \item{"Bacteria"} \item{"Protozoa"} \item{"Viruses"}
\item{"incertae sedis"} } It can also be one of the following not kingdoms:
#'\itemize{\item{Chordata} \item{Not Chordata}. Default: NULL.}}

\item{n_species}{integer The maximum number of species to return as examples
per pathway. For groups with less species than \code{n_species}, all
species are given. Default: 5.}

\item{kingdom}{character. Name of the column of \code{df} containing
information about kingdom. Default: \code{"kingdom"}.}
}
\value{
a data.frame
}
\description{
Function to get number of taxa introduced by different pathways. Possible
breakpoint: kingdom.
}
\examples{
library(readr)
data <- read_tsv("https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv")
get_table_pathways(data)
# Specify kingdom
get_table_pathways(data, "Plantae")
# Sepcify number of species to include in examples
get_table_pathways(data, "Plantae", n_species = 8)
}
