% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_distribution.R
\name{has_distribution}
\alias{has_distribution}
\title{Compare desired distribution information with actual one.
This function compares distribution information based on a single taxon key with 
user requests and returns a logical (TRUE or FALSE).
It is a function built on rgbif function `name_usage()`.}
\usage{
has_distribution(taxon_key, ...)
}
\arguments{
\item{taxon_key}{(single numeric or character) a single taxon key.}

\item{...}{one or more distribution properties and related values listed by GBIF:
http://rs.gbif.org/extension/gbif/1.0/distribution.xml.}
}
\value{
A data.frame with taxon_keys as first column and a second column 
(logical: TRUE or FALSE) based on match with user arguments
}
\description{
Compare desired distribution information with actual one.
This function compares distribution information based on a single taxon key with 
user requests and returns a logical (TRUE or FALSE).
It is a function built on rgbif function `name_usage()`.
}
