% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_introduction_year.R
\name{indicator_introduction_year}
\alias{indicator_introduction_year}
\title{Plot number of new introductions per year.}
\usage{
indicator_introduction_year(df, start_year_plot = 1920, smooth_span = 0.85,
  x_major_scale_stepsize = 10, x_minor_scale_stepsize = 5,
  facet_column = NULL)
}
\arguments{
\item{start_year_plot}{Year where the plot starts from.}

\item{smooth_span}{(numeric) Parameter for the applied Loess smoother. For
more information on the appropriate value, see \code{\link[stats]{loess}}.}

\item{facet_column}{NULL or character. The column to use to create additional
facet wrap plots underneath the main graph. When NULL, no facet graph are
created.}

\item{data}{A dataframe.}

\item{x_scale_stepsize}{(numeric) Parameter that indicates the major breaks
of the x axis}
}
\value{
A ggplot2 object.
}
\description{
Calculate how many new species has been introduced in a year.
}
\examples{
\dontrun{
indicator_introduction_year(data)
indicator_introduction_year(data, start_year_plot = 1940,
                            smooth_span = 0.6)
}
}
