% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_native_range_year.R
\name{indicator_native_range_year}
\alias{indicator_native_range_year}
\title{Create an interactive plot for the number of alien species per native region
and year of introduction}
\usage{
indicator_native_range_year(
  data,
  years = NULL,
  type = c("native_continent", "native_range"),
  x_lab = "year",
  y_lab = "alien species",
  relative = FALSE,
  first_observed = "first_observed"
)
}
\arguments{
\item{data}{input data.frame}

\item{years}{(numeric) vector years we are interested to. If \code{NULL}
(default) all years from minimum and maximum of years of first observation
are taken into account.}

\item{type}{character, native_range level of interest should be one of
\code{c("native_continent", "native_range")}}

\item{relative}{(logical) if TRUE, each bar is standardised before stacking}

\item{xlab}{character string, label of the x-axis. Default: "year".}

\item{ylab}{character string, label of the y-axis. Default: "number of alien
species".}
}
\value{
list with: \itemize{ \item{'static_plot': }{ggplot object, for a
given species the observed number per year and per native range is plotted
in a stacked bar chart} \item{'interactive_plot': }{plotly object, for a
given species the observed number per year and per native range is plotted
in a stacked bar chart} \item{'data': }{data displayed in the plot, as
data.frame with: \itemize{ \item{'year': }{year at which the species were
introduced} \item{'native_range': }{native range of the introduced species}
\item{'n': }{number of species introduced from the native range for a given
year} \item{'total': }{total number of species, from all around the world,
introduced during a given year} \item{'perc': }{percentage of species
introduced from the native range for a given year. (n/total)*100} } } }
}
\description{
Based on
\link[=https://github.com/inbo/reporting-rshiny-grofwildjacht/blob/exoten/reporting-grofwild/R/countYearProvince.R]{countYearProvince}
plot from reporting - rshiny - grofwildjacht
}
