% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_native_range_year.R
\name{indicator_native_range_year}
\alias{indicator_native_range_year}
\title{Create an interactive plot for the number of alien species per native region
and year of introduction}
\usage{
indicator_native_range_year(
  df,
  years = NULL,
  type = c("native_range", "native_continent"),
  x_major_scale_stepsize = 10,
  x_include_missing = FALSE,
  x_lab = "year",
  y_lab = "alien species",
  response_type = c("absolute", "relative", "cumulative"),
  relative = lifecycle::deprecated(),
  taxon_key_col = "key",
  first_observed = "first_observed"
)
}
\arguments{
\item{df}{input data.frame.}

\item{years}{(numeric) vector years we are interested to. If \code{NULL}
(default) all years from minimum and maximum of years of first observation
are taken into account.}

\item{type}{character, native_range level of interest should be one of
\code{c("native_range", "native_continent")}. Default: \code{"native_range"}. A
column called as the selected \code{type} must be present in \code{df}.}

\item{x_major_scale_stepsize}{(integer) Parameter that indicates the breaks
of the x axis. Default: 10.}

\item{x_include_missing}{(logical) if \code{TRUE} all consecutive years are
displayed on the x-axis, even if 0 records are available. If \code{FALSE}
(default) years with 0 count will be omitted and the x-axis is compressed.
Range is determined by either \code{years} if specified,
otherwise by the range of \code{first_observed} column in the \code{df}.}

\item{x_lab}{character string, label of the x-axis. Default: "year".}

\item{y_lab}{character string, label of the y-axis. Default: "number of alien
species".}

\item{response_type}{(character) summary type of response to be displayed;
should be one of \code{c("absolute", "relative", "cumulative")}.
Default: \code{"absolute"}. If "absolute" the number per year and location
is displayed; if "relative" each bar is standardised per year before stacking;
if "cumulative" the cumulative number over years per location.}

\item{relative}{(logical) if \code{TRUE} each bar is standardised before
stacking. Deprecated, use \code{response_type = "relative"} instead.}

\item{taxon_key_col}{character. Name of the column of \code{df} containing
taxon IDs. Default: \code{"key"}.}

\item{first_observed}{(character) Name of the column in \code{data}
containing temporal information about introduction of the alien species.
Expressed as years.}
}
\value{
list with:
\itemize{
\item \code{static_plot}: ggplot object, for a
given species the observed number per year and per native range is plotted
in a stacked bar chart.
\item \code{interactive_plot}: plotly object, for a
given species the observed number per year and per native range is plotted
in a stacked bar chart.
\item \code{data}: data displayed in the plot, as a data.frame with:
\itemize{
\item \code{year}: year at which the species were introduced.
\item \code{native_range}: native range of the introduced species.
\item \code{n}: number of species introduced from the native range for a given year.
\item \code{total}: total number of species, from all around the world, introduced.
during a given year.
\item \code{perc}: percentage of species introduced from the native range for a
given year, \code{n}/\code{total}*100.
}
}
}
\description{
Based on
\href{https://github.com/inbo/reporting-rshiny-grofwildjacht/blob/exoten/reporting-grofwild/R/countYearProvince.R}{countYearProvince}
plot from reporting - rshiny - grofwildjacht
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/indicators/master/data/",
  "interim/data_input_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
indicator_native_range_year(data, "native_continent", years = c(2010,2013))
}
}
