% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_pathway.R
\name{indicator_pathway}
\alias{indicator_pathway}
\title{Pathways of introduction}
\usage{
indicator_pathway(data, subcategory = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{subcategory}{(boolean) If TRUE, the subcategory levels are included in
the visualisation.}
}
\value{
A ggplot2 figure.
}
\description{
Make a figure of pathways of introduction. Use
  \code{subcategory} option to take into account only main categories or
  include subcategories as well.
}
\examples{
\dontrun{
library(readr)
datafile <- "https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv"
data <- read_tsv(datafile)
indicator_pathway(data)
# only main categories
indicator_pathway(data, subcategory = FALSE)
}
}
