% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_total_year.R
\name{indicator_total_year}
\alias{indicator_total_year}
\title{Create total invasive species indicator plot (TrIAS)}
\usage{
indicator_total_year(df, start_year_plot = 1940,
  x_major_scale_stepsize = 10, x_minor_scale_stepsize = 5,
  facet_column = NULL)
}
\arguments{
\item{df}{data.frame Contains the data as produced by the Trias pipeline,
with minimal columns.}

\item{start_year_plot}{int Limit to use as start year of the plot. For
scientific usage, the entire period could be relevant, but for policy
purpose, focusing on a more recent period could be required.}

\item{x_major_scale_stepsize}{int On which year interval labels are placed on
the x axis.}

\item{x_minor_scale_stepsize}{int On which year interval minor breaks are
placed on the x axis.}
}
\value{
ggplot2 object
}
\description{
Create total invasive species indicator plot (TrIAS)
}
\examples{
#' \dontrun{
# Load data
library(readr)
data <- read_tsv("https://raw.githubusercontent.com/trias-project/pipeline/master/data/interim/test_data_output_checklist_indicators.tsv")
start_year_plot <- 1900
x_major_scale_stepsize <- 25
x_minor_scale_stepsize <- 5
# without facets
indicator_total_year(data, start_year_plot, x_major_scale_stepsize)
# with facets
indicator_total_year(data, start_year_plot, x_major_scale_stepsize,
  x_minor_scale_stepsize, facet_column = "phylum")
  }
}
