% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_total_year.R
\name{indicator_total_year}
\alias{indicator_total_year}
\title{Create cumulative number of alien species indicator plot.}
\usage{
indicator_total_year(
  df,
  start_year_plot = 1940,
  x_major_scale_stepsize = 10,
  x_minor_scale_stepsize = 5,
  facet_column = NULL,
  first_observed = "first_observed",
  x_lab = "Year",
  y_lab = "Cumulative number of alien species"
)
}
\arguments{
\item{df}{df. Contains the data as produced by the Trias pipeline,
with minimal columns.}

\item{start_year_plot}{integer. Limit to use as start year of the plot. For
scientific usage, the entire period could be relevant, but for policy
purpose, focusing on a more recent period could be required.}

\item{x_major_scale_stepsize}{integer. On which year interval labels are
placed on the x axis.}

\item{x_minor_scale_stepsize}{integer. On which year interval minor breaks
are placed on the x axis.}

\item{facet_column}{NULL or character. Name of the column to use to create
additional facet wrap plots underneath the main graph. When NULL, no facet
graph is included. It is typically one of the highest taxonomic ranks, e.g.
\code{"kingdom"}, \code{"phylum"}, \code{"class"}, \code{"order"},
\code{"family"}. Other typical breakwdowns could be geographically related,
e.g. \code{"country"}, \code{"locality"}, \code{"pathway"} of
introduction or \code{"habitat"}.}

\item{first_observed}{character. Name of the column of \code{df} containing
information about year of introduction. Default: \code{"first_observed"}.}

\item{x_lab}{NULL or character. to set or remove the x-axis label.}

\item{y_lab}{NULL or character. to set or remove the y-axis label.}
}
\value{
ggplot2 object (or egg object if facets are used).
}
\description{
This function calculates the cumulative number of taxa introduced per
year. To do this, a column of input dataframe containing temporal information
about year of introduction is required.
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/pipeline/master/data/",
  "interim/test_data_output_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "NA",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    acceptedKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
start_year_plot <- 1900
x_major_scale_stepsize <- 25
x_minor_scale_stepsize <- 5
# without facets
indicator_total_year(data, start_year_plot, x_major_scale_stepsize)
# with facets
indicator_total_year(data, start_year_plot, facet_column = "phylum")
# specify name of column containing year of introduction (first_observed)
indicator_total_year(data, first_observed = "first_observed")
# specify axis labels
indicator_total_year(data, x_lab = "YEAR", y_lab = NULL)
}
}
