% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pathways_cbd.R
\docType{data}
\name{pathways_cbd}
\alias{pathways_cbd}
\title{Pathways of introduction as defined by the Convention on Biological Diversity}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 2 columns.
}
\usage{
pathways_cbd
}
\description{
This data frame contains all pathways of introdution at level 1
(\code{pathway_level1}) and level 2 (\code{pathway_level2}) as defined by the
Convention on Biological Diversity (CBD). Added pathway \code{unknown} at level 1
and level 2 for classifying taxa without pathway (at level 1 or level 2)
information.
}
\concept{pathways data}
\keyword{datasets}
