% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_cbd.R
\name{pathways_cbd}
\alias{pathways_cbd}
\title{Pathways of introduction as defined by CBD}
\usage{
pathways_cbd()
}
\value{
A tibble data.frame with 2 columns: \code{pathway_level1} and
\code{pathway_level2}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

It is deprecated as of trias 3.1.3. Please use the data frame
\url{pathwayscbd} directly instead.
}
\seealso{
Other checklist functions: 
\code{\link{gbif_get_taxa}()},
\code{\link{gbif_has_distribution}()},
\code{\link{gbif_verify_keys}()},
\code{\link{get_table_pathways}()},
\code{\link{indicator_introduction_year}()},
\code{\link{indicator_native_range_year}()},
\code{\link{indicator_total_year}()},
\code{\link{verify_taxa}()},
\code{\link{visualize_pathways_level1}()},
\code{\link{visualize_pathways_level2}()},
\code{\link{visualize_pathways_year_level1}()},
\code{\link{visualize_pathways_year_level2}()}
}
\concept{checklist functions}
