% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_cbd.R
\name{pathways_cbd}
\alias{pathways_cbd}
\title{Pathways of introduction as defined by CBD}
\usage{
pathways_cbd()
}
\value{
A tibble data.frame with 2 columns: \code{pathway_level1} and
\code{pathway_level2}.
}
\description{
Function to get all CBD pathays of introdution at level 1 (\code{pathway_level1})
and level 2 (\code{pathway_level2}). Added pathway \code{unknown} at level 1 and level
2 for classifying taxa without pathway (at level 1 or level 2) information.
}
