% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pathways_cbd.R, R/pathways_cbd.R
\docType{data}
\name{pathways_cbd}
\alias{pathways_cbd}
\title{Pathways of introduction as defined by CBD}
\format{
A tibble with 51 rows and 2 variables:
\itemize{
\item \code{pathway_level1}: (character) CBD pathway level 1
\item \code{pathway_level2}: (character) CBD pathway level 2
}
}
\source{
\href{https://www.cbd.int/}{CBD Standard}
}
\usage{
pathways_cbd

pathways_cbd()
}
\value{
A tibble data.frame with 2 columns: \code{pathway_level1} and
\code{pathway_level2}.
}
\description{
A tibble containing all CBD pathways of introduction at level 1
(\code{pathway_level1}) and level 2 (\code{pathway_level2}). Added pathway \code{unknown}
at level 1 and level 2 for classifying taxa without pathway (at level 1 or
level 2) information.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

It is deprecated as of trias 3.1.3. Please use the data frame
\url{pathways_cbd} directly. instead.
}
\concept{checklist data}
\keyword{datasets}
