% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_cbd.R
\name{pathways_cbd}
\alias{pathways_cbd}
\title{Pathways of introduction as defined by CBD}
\usage{
pathways_cbd()
}
\value{
A tibble data.frame with 2 columns: \code{pathway_level1} and
\code{pathway_level2}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

It is deprecated as of trias 3.1.3. Please use the data frame
\url{pathwayscbd} directly instead.
}
