% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_download_list.R
\name{update_download_list}
\alias{update_download_list}
\title{Update the list with all downloads}
\usage{
update_download_list(file, download_to_add, input_checklist, countries)
}
\arguments{
\item{file}{file containing all occurrence downloads from GBIF}

\item{download_to_add}{GBIF download key to be added in database}

\item{input_checklist}{input file with species}

\item{countries}{vector with countries}
}
\value{
message with the performed updates
}
\description{
This function opens a csv file containing all occurrence downloads from GBIF.
It updates the stauts of all downloads with status RUNNING.
In case download key is passed which is not present in the file it will be added as a new line.
}
