% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_download_list.R
\name{update_download_list}
\alias{update_download_list}
\title{Update the list with all downloads}
\usage{
update_download_list(
  file,
  download_to_add,
  input_checklist,
  url_doi_base = "https://doi.org/"
)
}
\arguments{
\item{file}{text file (tab separated) containing all occurrence downloads
from GBIF. File should contain the following columns:
\itemize{
\item \code{gbif_download_key}: GBIF download keys, e.g. 0012078-181003121212138
\item \code{input_checklist}: filename or link to readable text file containing the
list of queried taxa
\item \code{gbif_download_created}: datetime of the GBIF download creation as
returned by \code{rgbif} function  \code{occ_download_meta}
\item \code{gbif_download_status}: status of the GBIF download as returned by as
returned by \code{rgbif} function  \code{occ_download_meta}, e.g. \code{RUNNING},
\code{PREPARING}, \code{SUCCEDEED}
\item \code{gbif_download_doi}: DOI link as returned by as returned by \code{rgbif}
function  \code{occ_download_meta} (e.g. \code{10.15468/dl.kg0uxf}) + \code{url_doi_base}
value (see below)
}}

\item{download_to_add}{character. A GBIF download key to be added to file.}

\item{input_checklist}{text file with taxon keys whose occurrences you want
to download}

\item{url_doi_base}{character. doi base URL; \code{url_doi_base} + doi form a
link to a page with download information. Default: "https://doi.org/".}
}
\value{
message with the performed updates
}
\description{
This function opens a (tab-separated) text file containing all occurrence
downloads from GBIF and updates the status of all downloads with status
\code{RUNNING} or \code{PREPARING}. If the specifid download is not present it will be add.
}
\details{
If a download key is passed which is not present in the file it will be added
as a new line.
}
