% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_download_list.R
\name{update_download_list}
\alias{update_download_list}
\title{Update the list with all downloads}
\usage{
update_download_list(file, download_to_add, input_checklist,
  url_doi_base = "")
}
\arguments{
\item{file}{file containing all occurrence downloads from GBIF}

\item{download_to_add}{GBIF download key to be added to file}

\item{input_checklist}{file with taxon keys used for that download}

\item{url_doi_base}{(character) doi base URL; `url_doi_base` + doi form a link to a
page with download information; 
default: "https://doi.org/"}
}
\value{
message with the performed updates
}
\description{
This function opens a csv file containing all occurrence downloads from GBIF.
It updates the status of all downloads with status RUNNING.
}
\details{
If a download key is passed which is not present in the file it will be added as a new line.
}
