% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_taxa.R
\name{verify_taxa}
\alias{verify_taxa}
\title{Check and update verified taxa}
\usage{
verify_taxa(taxa, verified_taxa)
}
\arguments{
\item{taxa:}{a dataframe with at least the following columns: \itemize{
\item{checklist_scientificName} {: scientific name as provided by GBIF}
\item{checklist_datasetKey} {: dataset key (a UUID) of the checklist the
taxon comes from.} \item{backbone_taxonKey} {: a GBIF backbone key.}
\item{backbone_scientificName} {: scientific name as provided by GBIF
backbone.} \item{backbone_taxonomicStatus} {taxonomic status as provided by
GBIF backbone.} \item{backbone_acceptedName} {: accepted name (in case of
synonyms) as provided by GBIF backbone.}  \item{backbone_acceptedKey} {:
accepted key as provided by GBIF backbone.} \item{backbone_issues} {:
issues as provided by GBIF backbone.} }}

\item{verified_taxa:}{a dataframe with at least the following columns:
\itemize{ \item{checklist_scientificName} \item{backbone_scientificName}
\item{backbone_acceptedName} \item{backbone_taxonKey}
\item{backbone_acceptedKey} \item{verified_key}{: to be populated manually
by expert (not required by this function, but any other functionality will
use this key so it is good to check its existence)} \item{backbone_kingdom}
\item{date_added}{: to be populated by function} \item{issues}
\item{checklistst} {: checklists name where this taxa shows up},
\item{remarks} {: (optional) remarks as provided by the expert}}}
}
\value{
a list of five dataframes: \itemize{ \item{verified_taxa}{: same
  dataframe as input verified_taxa, but now with updated info.}
  \item{new_synonyms}{: a subset of verified_taxa (same columns) with
  added synonym relations (found in taxa, but not in verified_taxa)}
  \item{unused_synonyms}{: a subset of verified_taxa (same columns) with
  unused synonym relations (found in verified_taxa, but not in taxa)}
  \item{updated_scientificName}{: a dataframe with backbone_scientificName +
  updated_backbone_scientificName} \item{updated_accepted}{: a dataframe with
  backbone_accepted + updated_backbone_accepted} }
}
\description{
This function checks all verified taxa, add new synonyms or taxa not matched
to GBIF backbone in order to be evaluated by an expert, update taxa names in
case they have been changed and report the changes.
}
\examples{
taxa_in <- data.frame(
  checklist_scientificName = c("Aspius aspius",
                               "Rana catesbeiana",
                               "Polystichum tsus-simense J.Smith",
                               "Apus apus (Linnaeus, 1758)",
                               "Acmella spec."),
  checklist_datasetKey = c("98940a79-2bf1-46e6-afd6-ba2e85a26f9f",
                           "e4746398-f7c4-47a1-a474-ae80a4f18e92", 
                           "9ff7d317-609b-4c08-bd86-3bc404b77c42",
                           "39653f3e-8d6b-4a94-a202-859359c164c5",
                           "9ff7d317-609b-4c08-bd86-3bc404b77c42"),
  backbone_taxonKey = c(2360181, 2427092, 2651108, 5228676, NA),
  backbone_scientificName = c("Aspius aspius (Linnaeus, 1758)",
                              "Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-simense (Hook.) J.Sm.",
                              "Apus apus (Linnaeus, 1758)",
                              NA),
  backbone_kingdom = c("Animalia", "Animalia", "Plantae", "Plantae", NA),
  backbone_taxonomicStatus = c("SYNONYM", "SYNONYM", "SYNONYM", 
                               "ACCEPTED", NA),
  backbone_acceptedKey = c(5851603, 2427091, 4046493, NA, NA),
  backbone_acceptedName = c("Leuciscus aspius (Linnaeus, 1758)",
                            "Lithobates catesbeianus (Shaw, 1802)",
                            "Polystichum luctuosum (Kunze) Moore.",
                            NA, 
                            NA),
  stringsAsFactors = FALSE)
verified_taxa_in <- data.frame(
  checklist_scientificName = c("Rana catesbeiana",
                               "Polystichum tsus-simense J.Smith",
                               "Lemnaceae"),
  backbone_scientificName = c("Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-tsus-tsus (Hook.) Captain",
                              "Lemnaceae"),
  backbone_taxonomicStatus = c("SYNONYM", "SYNONYM", "SYNONYM"),
  backbone_acceptedName = c("Lithobates dummyus (Batman, 2018)",
                            "Polystichum luctuosum (Kunze) Moore.",
                            "Araceae"),
  backbone_taxonKey = c(2427092,
                        2651108,
                        6723),
  backbone_acceptedKey = c(2427091,
                          4046493,
                          6979),
  verified_key = c(2427091,
                   4046493,
                   6979),
  backbone_kingdom = c("Animalia", "Plantae", "Plantae"),
  date_added = as.Date(c("2018-07-01",
                         "2018-07-01",
                         "2018-07-01")),
  issues = c(NA, "ORIGINAL_NAME_DERIVED", NA),
  checklists = c("e4746398-f7c4-47a1-a474-ae80a4f18e92", 
                 "9ff7d317-609b-4c08-bd86-3bc404b77c42",
                 "e4746398-f7c4-47a1-a474-ae80a4f18e92"),
  remarks = c("dummy example 1: backbone_accepted should be updated",
              "dummy example 2: backbone_scientificName should be updated",
              "dummy example 3: nothing should be changed"),
  stringsAsFactors = FALSE)
verify_taxa(taxa = taxa_in, verified_taxa = verified_taxa_in)
}
