% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_taxa.R
\name{verify_taxa}
\alias{verify_taxa}
\title{Check and update verified taxa}
\usage{
verify_taxa(taxa, verified_taxa)
}
\arguments{
\item{taxa:}{a dataframe with at least the following columns: \itemize{
\item{checklist_scientificName} {: scientific name as provided by GBIF}
\item{checklist_datasetKey} {: dataset key (a UUID) of the checklist the
taxon comes from.} \item{backbone_taxonKey} {: a GBIF backbone key.}
\item{backbone_scientificName} {: scientific name as provided by GBIF
backbone.} \item{backbone_taxonomicStatus} {taxonomic status as provided by
GBIF backbone.} \item{backbone_acceptedName} {: accepted name (in case of
synonyms) as provided by GBIF backbone.}  \item{backbone_acceptedKey} {:
accepted key as provided by GBIF backbone.} \item{backbone_kingdom}
\item{backbone_issues} {: issues as provided by GBIF backbone.} }}

\item{verified_taxa:}{a dataframe with at least the following columns:
\itemize{ \item{checklist_scientificName} \item{backbone_scientificName}
\item{backbone_acceptedName} \item{backbone_taxonKey}
\item{backbone_acceptedKey} \item{verified_key}{: to be populated manually
by expert (not required by this function, but any other functionality will
use this key so it is good to check its existence)} \item{backbone_kingdom}
\item{date_added}{: to be populated by function} \item{issues}
\item{checklistst} {: checklists name where this taxa shows up},
\item{remarks} {: (optional) remarks as provided by the expert}}}
}
\value{
a list of five dataframes: \itemize{ \item{verified_taxa}{: same
  dataframe as input verified_taxa, but now with updated info.}
  \item{new_synonyms}{: a subset of verified_taxa (same columns) with added
  synonym relations (found in taxa, but not in verified_taxa)}
  \item{unused_taxa}{: a subset of verified_taxa (same columns) with
  unused taxa (found in verified_taxa, but not in taxa)}
  \item{updated_scientificName}{: a dataframe with backbone_scientificName +
  updated_backbone_scientificName} \item{updated_acceptedName}{: a dataframe
  with backbone_acceptedName + updated_backbone_acceptedName}
  \item{duplicates_taxa}{: a dataframe with all taxa present in more than one
  checklist.} }
}
\description{
This function checks all verified taxa, add new synonyms or taxa not matched
to GBIF backbone in order to be evaluated by an expert, update taxa names in
case they have been changed and report the changes.
}
\examples{
taxa_in <- data.frame(
  checklist_scientificName = c("Aspius aspius",
                               "Rana catesbeiana",
                               "Polystichum tsus-simense J.Smith",
                               "Apus apus (Linnaeus, 1758)",
                               "Begonia x semperflorens hort.",
                               "Rana catesbeiana",
                               "Spiranthes cernua (L.) Richard x S. odorata (Nuttall) Lindley"),
  checklist_datasetKey = c("98940a79-2bf1-46e6-afd6-ba2e85a26f9f",
                           "e4746398-f7c4-47a1-a474-ae80a4f18e92",
                           "9ff7d317-609b-4c08-bd86-3bc404b77c42",
                           "39653f3e-8d6b-4a94-a202-859359c164c5",
                           "9ff7d317-609b-4c08-bd86-3bc404b77c42",
                           "b351a324-77c4-41c9-a909-f30f77268bc4",
                           "9ff7d317-609b-4c08-bd86-3bc404b77c42"),
  backbone_taxonKey = c(2360181, 2427092, 2651108, 5228676, NA, 2427092, NA),
  backbone_scientificName = c("Aspius aspius (Linnaeus, 1758)",
                              "Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-simense (Hook.) J.Sm.",
                              "Apus apus (Linnaeus, 1758)",
                              NA,
                              "Rana catesbeiana Shaw, 1802",
                              NA),
  backbone_kingdom = c("Animalia", "Animalia", "Plantae", 
                       "Plantae", NA, "Animalia", NA),
  backbone_taxonomicStatus = c("SYNONYM", "SYNONYM", "SYNONYM",
                               "ACCEPTED", NA, "SYNONYM", NA),
  backbone_acceptedKey = c(5851603, 2427091, 4046493, NA, NA, 2427091, NA),
  backbone_acceptedName = c("Leuciscus aspius (Linnaeus, 1758)",
                            "Lithobates catesbeianus (Shaw, 1802)",
                            "Polystichum luctuosum (Kunze) Moore.",
                            NA, NA,
                            "Lithobates catesbeianus (Shaw, 1802)",
                            NA),
  backbone_issues = c("ORIGINAL_NAME_DERIVED", NA,
                      "ORIGINAL_NAME_DERIVED", NA, NA, NA, NA),
  stringsAsFactors = FALSE)

verified_taxa_in <- data.frame(
  checklist_scientificName = c("Rana catesbeiana",
                               "Polystichum tsus-simense J.Smith",
                               "Lemnaceae",
                               "Spiranthes cernua (L.) Richard x S. odorata (Nuttall) Lindley"),
  backbone_scientificName = c("Rana catesbeiana Shaw, 1802",
                              "Polystichum tsus-tsus-tsus (Hook.) Captain",
                              "Lemnaceae",
                              NA),
  backbone_taxonomicStatus = c("SYNONYM", "SYNONYM", "SYNONYM", NA),
  backbone_acceptedName = c("Lithobates dummyus (Batman, 2018)",
                            "Polystichum luctuosum (Kunze) Moore.",
                            "Araceae",
                            NA),
  backbone_taxonKey = c(2427092, 2651108, 6723,NA),
  backbone_acceptedKey = c(2427091, 4046493, 6979, NA),
  verified_key = c(2427091,
                   4046493,
                   6979,
                   "2805420,2805363"),
  backbone_kingdom = c("Animalia", "Plantae", "Plantae", NA),
  date_added = as.Date(c("2018-07-01",
                         "2018-07-01",
                         "2018-07-01",
                         "2018-07-16")),
  backbone_issues = c(NA_character_, NA_character_, NA_character_,
                      NA_character_),
  checklists = c("e4746398-f7c4-47a1-a474-ae80a4f18e92",
                 "9ff7d317-609b-4c08-bd86-3bc404b77c42",
                 "e4746398-f7c4-47a1-a474-ae80a4f18e92,39653f3e-8d6b-4a94-a202-859359c164c5",
                 "9ff7d317-609b-4c08-bd86-3bc404b77c42"),
  remarks = c("dummy example 1: backbone_acceptedName and checklists should be updated",
              "dummy example 2: backbone_scientificName and backbone_issues should be updated",
              "dummy example 3: nothing should be changed",
              "dummy example 4: multiple keys in verified_key are allowed"),
  stringsAsFactors = FALSE)
verify_taxa(taxa = taxa_in, verified_taxa = verified_taxa_in)
}
