% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pathways_level1.R
\name{visualize_pathways_level1}
\alias{visualize_pathways_level1}
\title{Plot number of introduced taxa for CBD pathways level 1}
\usage{
visualize_pathways_level1(
  df,
  category = NULL,
  from = NULL,
  facet_column = NULL,
  pathway_level1_names = "pathway_level1",
  taxon_names = "key",
  kingdom_names = "kingdom",
  phylum_names = "phylum",
  first_observed = "first_observed",
  cbd_standard = TRUE,
  title = NULL,
  x_lab = "Pathways",
  y_lab = "Number of introduced taxa"
)
}
\arguments{
\item{df}{df.}

\item{category}{NULL or character. One of the kingdoms as given in GBIF and
\code{Chordata} (the phylum), \verb{Not Chordata} (all other phyla of \code{Animalia}): 1.
\code{Plantae} 2. \code{Animalia} 3. \code{Fungi} 4. \code{Chromista} 5. \code{Archaea} 6.
\code{Bacteria} 7. \code{Protozoa} 8. \code{Viruses} 9. \verb{incertae sedis} 10. \code{Chordata}
11. \verb{Not Chordata} Default: \code{NULL}.}

\item{from}{NULL or numeric. Year trade-off: if not \code{NULL}
select only pathways related to taxa introduced during or after this year.
Default: \code{NULL}.}

\item{facet_column}{NULL or character. The column to use to create additional
facet wrap bar graphs underneath the main graph. When NULL, no facet graph
are created. One of \code{family}, \code{order}, \code{class}, \code{phylum}, \code{locality},
\code{native_range}, \code{habitat}. If column has another name, rename it before
calling this function. Default: \code{NULL}.}

\item{pathway_level1_names}{character. Name of the column of \code{df}
containing information about pathways at level 1. Default:
\code{pathway_level1}.}

\item{taxon_names}{character. Name of the column of \code{df} containing
information about taxa. This parameter is used to uniquely identify taxa.}

\item{kingdom_names}{character. Name of the column of \code{df} containing
information about kingdom. Default: \code{"kingdom"}.}

\item{phylum_names}{character. Name of the column of \code{df} containing
information about phylum. This parameter is used only if \code{category} is
one of:  \code{"Chordata"}, \code{"Not Chordata"}.  Default:
\code{"phylum"}.}

\item{first_observed}{character. Name of the column of \code{df} containing
information about year of introduction. Default: \code{"first_observed"}.}

\item{cbd_standard}{logical. If TRUE the values of pathway level 1 are
checked based on CBD standard as returned by \code{pathways_cbd()} and error is
returned if unmatched values are found. If FALSE, a warning is returned.
Default: TRUE.}

\item{title}{NULL or character. Title of the graph. Default: NULL.}

\item{x_lab}{NULL or character. Title of the graph. Default: "Pathways".}

\item{y_lab}{NULL or character. Title of the graph. Default: "Number of
introduced taxa".}
}
\value{
A egg object.
}
\description{
Function to plot bar graph with number of taxa introduced by different
pathways at level 1. Possible breakpoints: taxonomic (kingdoms +
vertebrates/invertebrates) and temporal (lower limit year).
}
\examples{
\dontrun{
library(readr)
datafile <- paste0(
  "https://raw.githubusercontent.com/trias-project/indicators/master/data/",
  "interim/data_input_checklist_indicators.tsv"
)
data <- read_tsv(datafile,
  na = "NA",
  col_types = cols(
    .default = col_character(),
    key = col_double(),
    nubKey = col_double(),
    speciesKey = col_double(),
    acceptedKey = col_double(),
    first_observed = col_double(),
    last_observed = col_double()
  )
)
# All taxa
visualize_pathways_level1(t)

# Animalia
visualize_pathways_level1(t, category = "Animalia")

# Chordata
visualize_pathways_level1(t, category = "Chordata")

# facet phylum
visualize_pathways_level1(t, category = "Animalia", facet_column = "phylum")

# facet habitat
visualize_pathways_level1(t, facet_column = "habitat")

# Only taxa introduced from 1950
visualize_pathways_level1(t, from = 1950)

# Add a title
visualize_pathways_level1(t, category = "Plantae", from = 1950, title = "Pathway level 1: Plantae. From 1950")

# Personalize axis labels
visualize_pathways_level1(t, x_lab = "Aantal taxa", y_lab = "pathways")
}
}
